Ensemble bug fixed and "working"
I put working in quotes because it doesn't actually classify very well.
With 300 relatively complex networks, the ensemble can only correctly
classify about 40% *of the training data*. I haven't even bothered
trying cross validation because I'm sure it will be pitiful. This seems
to be because the individual networks are not able to hit a 50% accuracy
rate. Perhaps a better system for generating the networks would be
helpful.